requirements.txt
fastapi
pydantic
uvicorn
mysql-connector-python
python-dotenv
passlib
jose
beautifulsoup4